by hjq

+++ HomeworkWeb/src/main/java/com/fiverings/homeworkweb/jparepository/UserJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.fiverings.homeworkweb.model.User;
 
 
-public interface UserJpaRepository extends JpaRepository<User, Integer>{
+public interface UserJpaRepository extends JpaRepository<User, Long>{
 
 }
